pyproject.toml
[project]
name = "weather-info"
version = "0.1.0"
description = "Python library to fetch weather data from multiple providers"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
weather = "weather.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["weather*"]
